+++ module-apps/application-phonebook/models/PhonebookModel.hpp
@@ -10,6 +10,7 @@
 #include "Interface/ContactRecord.hpp"
 #include "ListItemProvider.hpp"
 #include "NotesRecord.hpp"
+#include <module-apps/application-phonebook/widgets/PhonebookListView.hpp>
 
 #include <string>
 
@@ -18,14 +19,21 @@ class PhonebookModel : public app::DatabaseModel<ContactRecord>, public gui::Lis
   private:
     std::string queryFilter;
     std::uint32_t queryGroupFilter;
+    std::uint32_t queryDisplayMode;
+    std::uint32_t lastRequestedOffset = 0;
 
   public:
-    PhonebookModel(app::Application *app, std::string filter = "", std::uint32_t groupFilter = 0);
+    ContactsMapData letterMap;
+    PhonebookModel(app::Application *app,
+                   std::string filter        = "",
+                   std::uint32_t groupFilter = 0,
+                   std::uint32_t displayMode = 0);
     ~PhonebookModel() override = default;
 
     // virtual methods from DatabaseModel
     auto updateRecords(std::unique_ptr<std::vector<ContactRecord>> records) -> bool override;
     void requestRecords(const uint32_t offset, const uint32_t limit) override;
+    auto requestLetterMap() -> ContactsMapData;
 
     // virtual methods for ListViewProvider
     [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int override;
@@ -36,6 +44,11 @@ class PhonebookModel : public app::DatabaseModel<ContactRecord>, public gui::Lis
     [[nodiscard]] auto requestRecordsCount() -> unsigned int override;
 
     [[nodiscard]] auto getFilter() const -> const std::string &;
+
+    void setDisplayMode(std::uint32_t displayMode);
+    auto getDisplayMode() -> std::uint32_t;
+    auto getLastRequestedOffset() -> std::uint32_t;
+    auto getLabelMarkerDisplayMode() -> LabelMarkerDisplayMode;
 
     // onClick callback to register
     std::function<bool(gui::PhonebookItem *item)> messagesSelectCallback = nullptr;