+++ module-db/queries/phonebook/QueryContactGet.hpp
@@ -3,6 +3,7 @@
 #include <queries/RecordQuery.hpp>
 #include <queries/Filter.hpp>
 #include <Interface/ContactRecord.hpp>
+#include <module-apps/application-phonebook/data/ContactsMap.hpp>
 
 #include <string>
 
@@ -12,7 +13,7 @@ namespace db::query
      * @brief ContactRecord read query, filtered by text.
      *
      */
-    class ContactGet : public RecordQuery, public TextFilter, public ContactGroupFilter
+    class ContactGet : public RecordQuery, public TextFilter, public ContactGroupFilter, public ContactDisplayMode
     {
       public:
         /**
@@ -20,7 +21,9 @@ namespace db::query
          *
          * @param filter text filter
          */
-        ContactGet(const std::string &filter = "", const std::uint32_t &groupFilter = 0);
+        ContactGet(const std::string &filter        = "",
+                   const std::uint32_t &groupFilter = 0,
+                   const std::uint32_t &displayMode = 0);
 
         /**
          * @brief Construct read query with limit, offset and filter values
@@ -32,7 +35,8 @@ namespace db::query
         ContactGet(std::size_t limit,
                    std::size_t offset,
                    const std::string &filter        = "",
-                   const std::uint32_t &groupFilter = 0);
+                   const std::uint32_t &groupFilter = 0,
+                   const std::uint32_t &displayMode = 0);
 
         /**
          * @brief debug info
@@ -69,7 +73,10 @@ namespace db::query
      * @brief A query to get a number of contacts filtered with a text
      *
      */
-    class ContactGetSize : public RecordsSizeQuery, public TextFilter, public ContactGroupFilter
+    class ContactGetSize : public RecordsSizeQuery,
+                           public TextFilter,
+                           public ContactGroupFilter,
+                           public ContactDisplayMode
     {
       public:
         /**
@@ -77,7 +84,9 @@ namespace db::query
          *
          * @param filter text to filter contacts with
          */
-        ContactGetSize(const std::string &filter = "", const std::uint32_t &groupFilter = 0);
+        ContactGetSize(const std::string &filter        = "",
+                       const std::uint32_t &groupFilter = 0,
+                       const std::uint32_t &displayMode = 0);
 
         /**
          * @brief debug info
@@ -85,6 +94,29 @@ namespace db::query
          * @return class name
          */
         [[nodiscard]] auto debugInfo() const -> std::string override;
+    };
+
+    class ContactGetLetterMap : public RecordQuery,
+                                public TextFilter,
+                                public ContactGroupFilter,
+                                public ContactDisplayMode
+    {
+      public:
+        ContactGetLetterMap(const std::string &filter        = "",
+                            const std::uint32_t &groupFilter = 0,
+                            const std::uint32_t &displayMode = 0);
+
+        ContactGetLetterMap(std::size_t limit,
+                            std::size_t offset,
+                            const std::string &filter        = "",
+                            const std::uint32_t &groupFilter = 0,
+                            const std::uint32_t &displayMode = 0);
+    };
+
+    class ContactGetLetterMapResult : public LetterMapResult
+    {
+      public:
+        ContactGetLetterMapResult(ContactsMapData &LetterMap);
     };
 
 }; // namespace db::query